修正行走被改快... 隨機調整 utils 的程式碼確認完畢
git-svn-id: http://www.castiello-tsai.org:8523/svn/L1JTW@155 96aee97a-5c11-ed42-8518-d7f33e96f5f9

+++ src/l1j/server/server/utils/ByteArrayUtil.java
@@ -28,7 +28,7 @@ public class ByteArrayUtil {
 
 	public String getTerminatedString(int i) {
 		StringBuffer stringbuffer = new StringBuffer();
-		for (int j = i; j < _byteArray.length && _byteArray[j] != 0; j++) {
+		for (short j = i; j < _byteArray.length && _byteArray[j] != 0; j++) {
 			stringbuffer.append((char) _byteArray[j]);
 		}
 
@@ -37,8 +37,8 @@ public class ByteArrayUtil {
 
 	public String dumpToString() {
 		StringBuffer stringbuffer = new StringBuffer();
-		int j = 0;
-		for (int k = 0; k < _byteArray.length; k++) {
+		short j = 0;
+		for (short k = 0; k < _byteArray.length; k++) {
 			if (j % 16 == 0) {
 				stringbuffer.append((new StringBuilder()).append(fillHex(k, 4))
 						.append(": ").toString());
@@ -49,8 +49,8 @@ public class ByteArrayUtil {
 				continue;
 			}
 			stringbuffer.append("   ");
-			int i1 = k - 15;
-			for (int l1 = 0; l1 < 16; l1++) {
+			short i1 = k - 15;
+			for (short l1 = 0; l1 < 16; l1++) {
 				byte byte0 = _byteArray[i1++];
 				if (byte0 > 31 && byte0 < 128) {
 					stringbuffer.append((char) byte0);
@@ -63,14 +63,14 @@ public class ByteArrayUtil {
 			j = 0;
 		}
 
-		int l = _byteArray.length % 16;
+		short l = _byteArray.length % 16;
 		if (l > 0) {
-			for (int j1 = 0; j1 < 17 - l; j1++) {
+			for (short j1 = 0; j1 < 17 - l; j1++) {
 				stringbuffer.append("   ");
 			}
 
-			int k1 = _byteArray.length - l;
-			for (int i2 = 0; i2 < l; i2++) {
+			short k1 = _byteArray.length - l;
+			for (short i2 = 0; i2 < l; i2++) {
 				byte byte1 = _byteArray[k1++];
 				if (byte1 > 31 && byte1 < 128) {
 					stringbuffer.append((char) byte1);
@@ -86,7 +86,7 @@ public class ByteArrayUtil {
 
 	private String fillHex(int i, int j) {
 		String s = Integer.toHexString(i);
-		for (int k = s.length(); k < j; k++) {
+		for (short k = s.length(); k < j; k++) {
 			s = (new StringBuilder()).append("0").append(s).toString();
 		}
